feat: 모달 컴포넌트 개선 및 애니메이션 추가
