+++ fe/src/app/components/shared/modal/Modal.tsx
@@ -1,49 +1,40 @@
 'use client';
 
-import { useEffect, useState, ReactNode } from 'react';
+import { useEffect, useState } from 'react';
 import { createPortal } from 'react-dom';
 import { modalStore } from './modal.store';
 import styles from './modal.module.css';
+import { ModalProps } from './type';
+import CSSUtil from '@/utils/css';
 
-export function useModal() {
-  const openModal = modalStore((state) => state.openModal);
-  const closeModal = modalStore((state) => state.closeModal);
-  const isOpen = modalStore((state) => state.isOpen);
-
-  return { openModal, closeModal, isOpen };
-}
-
-interface ModalProps {
-  id: string;
-  children: ReactNode;
-}
+type Event = React.MouseEvent<HTMLDivElement>;
 
 export default function Modal({ id, children }: ModalProps) {
   const openModals = modalStore((state) => state.openModals);
   const closeModal = modalStore((state) => state.closeModal);
   const [mounted, setMounted] = useState(false);
+  const [isVisible, setIsVisible] = useState(false);
 
-  useEffect(() => {
-    setMounted(true);
-  }, []);
-
-  if (!mounted) return null;
+  useEffect(() => setMounted(true), []);
 
   const isOpen = openModals.has(id);
-  if (!isOpen) return null;
 
-  const handleBackdropClick = (e: React.MouseEvent<HTMLDivElement>) => {
-    if (e.target === e.currentTarget) {
-      closeModal(id);
-    }
+  useEffect(() => {
+    if (isOpen && mounted) requestAnimationFrame(() => setIsVisible(true));
+    else setIsVisible(false);
+  }, [isOpen, mounted]);
+
+  if (!mounted || !isOpen) return null;
+
+  const handleContentClick = (e: Event) => e.stopPropagation();
+  const handleBackdropClick = (e: Event) => {
+    if (e.target === e.currentTarget) closeModal(id);
   };
 
-  const handleContentClick = (e: React.MouseEvent<HTMLDivElement>) => {
-    e.stopPropagation();
-  };
+  const backdropClassName = CSSUtil.buildCls(styles.backdrop, isVisible && styles.visible);
 
   return createPortal(
-    <div className={styles.backdrop} onClick={handleBackdropClick}>
+    <div className={backdropClassName} onClick={handleBackdropClick}>
       <div className={styles.content} onClick={handleContentClick}>
         {children}
       </div>
