+++ fe/src/app/components/shared/modal/modal.module.css
@@ -4,9 +4,15 @@
   left: 0;
   width: 100%;
   height: 100%;
-  background-color: #0000004d;
+  background-color: #00000066;
   display: flex;
   justify-content: center;
   align-items: center;
   z-index: 1000;
+  opacity: 0;
+  transition: opacity .2s ease-in-out;
+}
+
+.backdrop.visible {
+  opacity: 1;
 }
