+++ fe/src/app/components/shared/modal/type.ts
@@ -1,0 +1,6 @@
+import { ReactNode } from 'react';
+
+export interface ModalProps {
+  id: string;
+  children: ReactNode;
+}
