+++ fe/src/app/components/shared/modal/useModal.ts
@@ -1,0 +1,9 @@
+import { modalStore } from './modal.store';
+
+export function useModal() {
+  const openModal = modalStore((state) => state.openModal);
+  const closeModal = modalStore((state) => state.closeModal);
+  const isOpen = modalStore((state) => state.isOpen);
+
+  return { openModal, closeModal, isOpen };
+}
